requirements.txt
-i https://download.pytorch.org/whl/torch_stable.html

hydra-core
torch==1.9.0+cu111
transformers
numpy
